feat: Move hooks to common folder
