+++ src/DateContainer/index.tsx
@@ -1,4 +1,4 @@
-import useTimer from '../hooks/useTimer'
+import useTimer from '../common/hooks/useTimer'
 import DateList from './DateList'
 
 const DateContainer = () => {
